- Copy & Paste the elevator logic and use an own ui file for each game - Changed how teleporters are assigned to the dock connections
