+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [6.2.0] - 2023-09-??
 
 ### Metroid Dread
+- Added: Elevator and Shuttle randomizer.
 
 #### Logic Database
