+++ randovania/games/dread/generator/base_patches_factory.py
@@ -8,19 +8,22 @@ from randovania.game_description.requirements.base import Requirement
 from randovania.game_description.requirements.requirement_and import RequirementAnd
 from randovania.game_description.requirements.resource_requirement import ResourceRequirement
 from randovania.games.dread.layout.dread_configuration import DreadConfiguration
-from randovania.generator.base_patches_factory import ElevatorRandoBasePatchesFactory
+from randovania.generator.base_patches_factory import BasePatchesFactory
+from randovania.generator.elevator_distributor import get_dock_connections_for_elevators
 
 if TYPE_CHECKING:
-    from collections.abc import Iterator
+    from collections.abc import Iterable, Iterator
     from random import Random
 
     from randovania.game_description.assignment import NodeConfigurationAssociation
+    from randovania.game_description.db.dock_node import DockNode
+    from randovania.game_description.db.node import Node
     from randovania.game_description.game_description import GameDescription
     from randovania.game_description.game_patches import GamePatches
     from randovania.layout.base.base_configuration import BaseConfiguration
 
 
-class DreadBasePatchesFactory(ElevatorRandoBasePatchesFactory):
+class DreadBasePatchesFactory(BasePatchesFactory):
     def configurable_node_assignment(self, configuration: DreadConfiguration, game: GameDescription,
                                      rng: Random) -> Iterator[NodeConfigurationAssociation]:
         result = []
@@ -74,3 +77,8 @@ class DreadBasePatchesFactory(ElevatorRandoBasePatchesFactory):
             (game.region_list.node_by_identifier(identifier), target)
             for identifier, target in dock_weakness
         ))
+
+    def dock_connections_assignment(self, configuration: DreadConfiguration,
+                                    game: GameDescription, rng: Random ) -> Iterable[tuple[DockNode, Node]]:
+        dock_assignment = get_dock_connections_for_elevators(configuration, game, rng)
+        yield from dock_assignment
