+++ randovania/games/dread/gui/preset_settings/__init__.py
@@ -8,7 +8,7 @@ if TYPE_CHECKING:
 
 
 def dread_preset_tabs(editor: PresetEditor, window_manager: WindowManager):
-    from randovania.games.common.prime_family.gui.elevators_tab import PresetElevatorsDread
+    from randovania.games.dread.gui.preset_settings.dread_elevators_tab import PresetElevatorsDread
     from randovania.games.dread.gui.preset_settings.dread_energy_tab import PresetDreadEnergy
     from randovania.games.dread.gui.preset_settings.dread_generation_tab import PresetDreadGeneration
     from randovania.games.dread.gui.preset_settings.dread_goal_tab import PresetDreadGoal
