+++ randovania/games/prime1/gui/preset_settings/__init__.py
@@ -9,7 +9,7 @@ if TYPE_CHECKING:
 
 def prime1_preset_tabs(editor: PresetEditor, window_manager: WindowManager):
 
-    from randovania.games.common.prime_family.gui.elevators_tab import PresetElevatorsPrime1
+    from randovania.games.common.prime_family.gui.elevators_tab_prime1 import PresetElevatorsPrime1
     from randovania.games.prime1.gui.preset_settings.prime_enemy_stat_randomizer import PresetEnemyAttributeRandomizer
     from randovania.games.prime1.gui.preset_settings.prime_generation_tab import PresetPrimeGeneration
     from randovania.games.prime1.gui.preset_settings.prime_goal_tab import PresetPrimeGoal
