+++ randovania/games/prime1/layout/prime_configuration.py
@@ -5,11 +5,13 @@ from enum import Enum
 
 from randovania.bitpacking.bitpacking import BitPackDataclass, BitPackEnum
 from randovania.bitpacking.json_dataclass import JsonDataclass
+from randovania.games.common.prime_family.layout.lib.prime_trilogy_teleporters import (
+    PrimeTrilogyTeleporterConfiguration,
+)
 from randovania.games.game import RandovaniaGame
 from randovania.games.prime1.layout.artifact_mode import LayoutArtifactMode
 from randovania.games.prime1.layout.hint_configuration import HintConfiguration
 from randovania.layout.base.base_configuration import BaseConfiguration
-from randovania.layout.lib.teleporters import TeleporterConfiguration
 
 
 class RoomRandoMode(BitPackEnum, Enum):
@@ -43,7 +45,7 @@ class EnemyAttributeRandomizer(BitPackDataclass, JsonDataclass):
 
 @dataclasses.dataclass(frozen=True)
 class PrimeConfiguration(BaseConfiguration):
-    elevators: TeleporterConfiguration
+    elevators: PrimeTrilogyTeleporterConfiguration
     hints: HintConfiguration
     energy_per_tank: int = dataclasses.field(metadata={"min": 1, "max": 1000, "precision": 1})
     artifact_target: LayoutArtifactMode
