+++ randovania/games/prime2/generator/base_patches_factory.py
@@ -11,8 +11,10 @@ from randovania.game_description.requirements.requirement_and import Requirement
 from randovania.game_description.requirements.resource_requirement import ResourceRequirement
 from randovania.game_description.resources import search
 from randovania.game_description.resources.resource_type import ResourceType
+from randovania.games.prime2.layout.echoes_configuration import EchoesConfiguration
 from randovania.games.prime2.layout.translator_configuration import LayoutTranslatorRequirement
-from randovania.generator.base_patches_factory import ElevatorRandoBasePatchesFactory, MissingRng
+from randovania.generator.base_patches_factory import BasePatchesFactory, MissingRng
+from randovania.generator.elevator_distributor import get_dock_connections_for_elevators
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
@@ -20,9 +22,9 @@ if TYPE_CHECKING:
 
     from randovania.game_description.assignment import NodeConfigurationAssociation
     from randovania.game_description.db.dock import DockWeakness
+    from randovania.game_description.db.node import Node
     from randovania.game_description.game_description import GameDescription
-    from randovania.game_description.game_patches import ElevatorConnection, GamePatches
-    from randovania.games.prime2.layout.echoes_configuration import EchoesConfiguration
+    from randovania.game_description.game_patches import GamePatches
     from randovania.layout.base.base_configuration import BaseConfiguration
 
 
@@ -69,7 +71,7 @@ WORLDS = [
 ]
 
 
-class EchoesBasePatchesFactory(ElevatorRandoBasePatchesFactory):
+class EchoesBasePatchesFactory(BasePatchesFactory):
     def create_base_patches(self,
                             configuration: BaseConfiguration,
                             rng: Random,
@@ -77,19 +79,19 @@ class EchoesBasePatchesFactory(ElevatorRandoBasePatchesFactory):
                             is_multiworld: bool,
                             player_index: int,
                             rng_required: bool = True) -> GamePatches:
+        assert isinstance(configuration, EchoesConfiguration)
         parent = super().create_base_patches(configuration, rng, game, is_multiworld, player_index, rng_required)
         return self.assign_save_door_weaknesses(parent, configuration, game)
 
-    def add_elevator_connections_to_patches(self, configuration: EchoesConfiguration, rng: Random,
-                                            patches: GamePatches) -> GamePatches:
-        patches = super().add_elevator_connections_to_patches(configuration, rng, patches)
+    def dock_connections_assignment(self, configuration: EchoesConfiguration,
+                                    game: GameDescription, rng: Random ) -> Iterable[tuple[DockNode, Node]]:
+        dock_assignment = get_dock_connections_for_elevators(configuration, game, rng)
 
         if not configuration.portal_rando:
-            return patches
+            yield from dock_assignment
+            return
 
-        dock_assignment = []
-
-        for world in patches.game.region_list.regions:
+        for world in game.region_list.regions:
             light_portals = []
             dark_portals = []
 
@@ -107,44 +109,7 @@ class EchoesBasePatchesFactory(ElevatorRandoBasePatchesFactory):
             dock_assignment.extend(zip(light_portals, dark_portals))
             dock_assignment.extend(zip(dark_portals, light_portals))
 
-        return patches.assign_dock_connections(dock_assignment)
-
-    def elevator_echoes_shuffled(self, configuration: EchoesConfiguration, patches: GamePatches,
-                                 rng: Random) -> ElevatorConnection:
-        worlds = list(WORLDS)
-        rng.shuffle(worlds)
-
-        result = {}
-
-        def area_to_node(identifier: AreaIdentifier):
-            area = patches.game.region_list.area_by_area_location(identifier)
-            for node in area.actual_nodes:
-                if node.valid_starting_location:
-                    return node.identifier
-            raise KeyError(f"{identifier} has no valid starting location")
-
-        def link_to(source: AreaIdentifier, target: AreaIdentifier):
-            result[area_to_node(source)] = area_to_node(target)
-            result[area_to_node(target)] = area_to_node(source)
-
-        def tg_link_to(source: str, target: AreaIdentifier):
-            link_to(AreaIdentifier("Temple Grounds", source), target)
-
-        # TG -> GT
-        tg_link_to("Temple Transport A", worlds[0].front)
-        tg_link_to("Temple Transport B", worlds[0].left)
-        tg_link_to("Temple Transport C", worlds[0].right)
-
-        tg_link_to("Transport to Agon Wastes", worlds[1].front)
-        tg_link_to("Transport to Torvus Bog", worlds[2].front)
-        tg_link_to("Transport to Sanctuary Fortress", worlds[3].front)
-
-        # inter areas
-        link_to(worlds[1].right, worlds[2].left)
-        link_to(worlds[2].right, worlds[3].left)
-        link_to(worlds[3].right, worlds[1].left)
-
-        return result
+        yield from dock_assignment
 
     def configurable_node_assignment(self, configuration: EchoesConfiguration, game: GameDescription,
                                      rng: Random) -> Iterable[NodeConfigurationAssociation]:
