+++ test/gui/preset_settings/test_preset_elevator.py
@@ -9,22 +9,29 @@ from PySide6 import QtCore
 
 from randovania.game_description import default_database
 from randovania.game_description.db.area_identifier import AreaIdentifier
-from randovania.games.common.prime_family.gui.elevators_tab import PresetElevators
+from randovania.games.common.prime_family.gui.elevators_tab_prime1 import PresetElevatorsPrime1
+from randovania.games.common.prime_family.gui.elevators_tab_prime2 import PresetElevatorsPrime2
+from randovania.games.dread.gui.preset_settings.dread_elevators_tab import PresetElevatorsDread
 from randovania.games.game import RandovaniaGame
 from randovania.games.prime1.layout import prime_configuration
 from randovania.interface_common.preset_editor import PresetEditor
 from randovania.layout.lib.teleporters import TeleporterShuffleMode, TeleporterTargetList
 
 
-@pytest.mark.parametrize("game", [RandovaniaGame.METROID_PRIME, RandovaniaGame.METROID_PRIME_ECHOES])
+@pytest.mark.parametrize("game", [RandovaniaGame.METROID_PRIME,
+                                  RandovaniaGame.METROID_PRIME_ECHOES, RandovaniaGame.METROID_DREAD])
 def test_on_preset_changed(skip_qtbot, preset_manager, game):
     # Setup
     base = preset_manager.default_preset_for_game(game).get_preset()
     preset = dataclasses.replace(base, uuid=uuid.UUID('b41fde84-1f57-4b79-8cd6-3e5a78077fa6'))
     options = MagicMock()
     editor = PresetEditor(preset, options)
-    window = PresetElevators(editor, default_database.game_description_for(preset.game), MagicMock())
-
+    if game == RandovaniaGame.METROID_PRIME:
+        window = PresetElevatorsPrime1(editor, default_database.game_description_for(preset.game), MagicMock())
+    elif game == RandovaniaGame.METROID_PRIME_ECHOES:
+        window = PresetElevatorsPrime2(editor, default_database.game_description_for(preset.game), MagicMock())
+    else:
+        window = PresetElevatorsDread(editor, default_database.game_description_for(preset.game), MagicMock())
     # Run
     window.on_preset_changed(editor.create_custom_preset_with())
 
@@ -51,7 +58,7 @@ def test_check_credits(skip_qtbot, preset_manager):
     )
     options = MagicMock()
     editor = PresetEditor(preset, options)
-    window = PresetElevators(editor, default_database.game_description_for(preset.game), MagicMock())
+    window = PresetElevatorsPrime1(editor, default_database.game_description_for(preset.game), MagicMock())
     window.on_preset_changed(editor.create_custom_preset_with())
 
     # Run
