+++ test/test_files/log_files/dread/elevator_rando.rdvgame
@@ -200,9 +200,7 @@
                     "elevators": {
                         "mode": "one-way-elevator",
                         "excluded_teleporters": [],
-                        "excluded_targets": [],
-                        "skip_final_bosses": false,
-                        "allow_unvisited_room_names": false
+                        "excluded_targets": []
                     },
                     "energy_per_tank": 100,
                     "immediate_energy_parts": true,
